Add support for customizing GPT model via environment variable
This commit adds the ability to customize the GPT model used by the
program via the "GPT_MODEL" environment variable. If the variable is
not set, the program will default to using the GPT4 model. The model
is used in the createChatCompletion function, which has been updated
to use the "gptModel" field of the Frederica struct instead of the
hardcoded GPT4 model.